projects/scrapers/web_crawl/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "web-crawl"
version = "0.1.0"
description = "Professional web crawler with AI-powered content analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "google-generativeai>=0.8.0",
    "python-dotenv>=1.0.0",
    "urllib3>=2.0.0",
    "lxml>=4.9.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "aiohttp>=3.9.0",
    "markdownify>=0.13.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[project.scripts]
web-crawl = "crawl:main"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
